#50 refactor: position 参照元を新 _stores/position/ へ切替
- hook (usePositionStore/usePositionStoreApi) は新 _stores/position/ (実装/context) を参照
- component/context の DEFAULT_POSITION は _stores/position/constants へ直接参照
- intent-store.ts の PositionStore 型・DEFAULT_POSITION は _types/_constants 集約 index 経由に統一
  (旧 _stores/position-store.ts・_contexts/position-store-context.tsx 自体はまだ削除しない)
