+++ src/prototypes/time-control/time-control-03/__tests__/position-store.test.ts
@@ -2,7 +2,7 @@ import { createActorStore } from '../_stores/actor'
 import { createActorSettingsStore } from '../_stores/actor-settings'
 import { createGameClockStore } from '../_stores/game-clock'
 import { createPathStore } from '../_stores/path'
-import { createPositionStore } from '../_stores/position-store'
+import { createPositionStore } from '../_stores/position'
 
 beforeEach(() => {
   vi.useFakeTimers()
