+++ src/prototypes/time-control/time-control-03/_contexts/stores-provider.tsx
@@ -15,10 +15,9 @@ import {
   createPlannedPathStore,
   PlannedPathStoreContext,
 } from '../_stores/planned-path'
-import { createPositionStore } from '../_stores/position-store'
+import { createPositionStore, PositionStoreContext } from '../_stores/position'
 
 import { IntentStoreContext } from './intent-store-context'
-import { PositionStoreContext } from './position-store-context'
 
 type StoresProviderProps = {
   children: ReactNode
