+++ src/prototypes/time-control/time-control-03/_stores/intent-store.ts
@@ -4,15 +4,19 @@ import { getTickMs } from '../../time-control-02/_lib/get-tick-ms'
 import { generateMovePath } from '../_lib/generate-move-path'
 import { ActionLogEntry, MoveIntent } from '../types'
 
-import { DEFAULT_ACTOR_INFO, DEFAULT_ACTOR_SETTINGS } from './_constants'
+import {
+  DEFAULT_ACTOR_INFO,
+  DEFAULT_ACTOR_SETTINGS,
+  DEFAULT_POSITION,
+} from './_constants'
 import {
   ActorSettingsStore,
   ActorStore,
   GameClockStore,
   PathStore,
   PlannedPathStore,
+  PositionStore,
 } from './_types'
-import { DEFAULT_POSITION, PositionStore } from './position-store'
 
 /** 状態を持たない操作専用 store。移動企図 (経路生成 → path-store への書き込み) のみを行う */
 export type IntentState = {
@@ -27,7 +31,7 @@ export type IntentStore = StoreApi<IntentState>
  * @param actorSettingsStore 経路の最低 step 数 (`_stores/actor-settings/store.ts`)
  * @param pathStore 生成した経路の書き込み先 (`_stores/path/store.ts`)
  * @param plannedPathStore 予定位置表示用の経路の書き込み先 (`_stores/planned-path/store.ts`)
- * @param positionStore 企図元の現在位置 (`_stores/position-store.ts`)
+ * @param positionStore 企図元の現在位置 (`_stores/position/store.ts`)
  * @param gameClockStore 共通ゲームクロック・履歴ログ (`_stores/game-clock/store.ts`)
  */
 export const createIntentStore = (
